Improvement: Improve printing left bar settings
